extract function editor
